arkiv_sdk

Module node_bindings

Used for running a geth node with arkiv capabilities locally for development and testing purposes. Constructor semantics work similarly to alloy::node_bindings::Anvil and alloy::node_bindings::Geth.

Structs

Arkiv

Builder for launching a local geth node with arkiv capabilities.

ArkivInstance

An op-geth instance with arkiv capabilities. The instance is closed on ops::Drop::drop.

Enums

Tag

The release tag to fetch from GitHub.