arkiv_sdk::node_bindings

Struct Arkiv

pub struct Arkiv { /* private fields */ }

Builder for launching a local geth node with arkiv capabilities.

Example

Assuming a geth executable with arkiv capabilities is on $PATH, the following is equivalent to running

geth --dev \
  --networkid 1337 \
  --http --http.api 'eth,web3,net,debug,arkiv' \
  --http.addr '0.0.0.0' --http.port 8545 \
  --http.corsdomain '*' --http.vhosts '*' \
  --ws --ws.api 'eth,web3,net,debug,arkiv' \
  --ws.addr '0.0.0.0' --ws.port 8546 \
  --datadir './geth_data' --verbosity 3

Arkiv::fetch_tag can be used to fetch and run geth from the Arkiv-Network GitHub releases.

use arkiv_sdk::node_bindings::Arkiv;

let arkiv = Arkiv::default().spawn()?;

drop(arkiv); // kill the child process

Implementations

impl Arkiv

pub const GITHUB_RELEASE_URL: &str = "https://api.github.com/repos/Arkiv-Network/arkiv-op-geth/releases"

pub const PROGRAM: &str = "geth"

pub const ARKIV_NETWORKID: u64 = 1_337u64

This networkid is automatically set internally by the node on startup, however, we set it explicitly in Arkiv::default since this may produce a race condition for wallets, providers and txs when setting the chain_id field.

pub const API: &str = "eth,web3,net,debug,arkiv"

pub const DEFAULT_ADDR: &str = "localhost"

pub const DEFAULT_HTTP_PORT: u16 = 8_545u16

pub const DEFAULT_WS_PORT: u16 = 8_546u16

pub const NODE_STARTUP_TIMEOUT: Duration

pub fn new() -> Self

Construct an Arkiv builder with all options unset.

pub fn fetch_tag(self, tag: Tag) -> Self

Sets the option to fetch a prebuilt op-geth tagged release from https://github.com/Arkiv-Network/arkiv-op-geth/releases. The release is unpacked at the specified directory provided by Arkiv::download_dir, otherwise defaulting to $XDG_CONFIG_HOME/arkiv, appended by the tag. The download is not triggered until Arkiv::spawn, and will avoid needless network calls. Successive calls will not trigger re-downloads unless the tag directory is deleted.

NOTE: Arkiv::program will override this setting, which is useful when developing in offline mode or testing in sandboxed environments that cannot make network calls.

If the checksum cannot be verified, the fetched binary will not be executed and an error will be returned from Arkiv::spawn.

Errors

If the download directory does not exist, or if the final component of the path is not a directory. See fs::canonicalize for further details.

If the home config directory does not exist, or could not be created. See [dirs::config_dir] and fs::create_dir_all for further details.

pub fn download_dir<T: Into<PathBuf>>(self, download_dir: T) -> Self

Sets the directory to download an arkiv op-geth release.

Setting this option alone will have no effect. See Arkiv::fetch_tag for details.

pub fn program<T: Into<PathBuf>>(self, path: T) -> Self

Overrides the path to the geth program.

pub fn dev(self) -> Self

Whether to launch the geth instance in --dev mode.

pub fn dev_period(self, period: Duration) -> Self

Block period to use in developer mode (0 = mine only if transaction pending)

pub fn http(self) -> Self

Whether to pass the --http flag to the geth instance.

pub fn http_api<T: Into<String>>(self, api: T) -> Self

The --http.api which will be used when the geth instance is launched.

pub fn http_addr<T: Into<String>>(self, addr: T) -> Self

Sets the --http.addr which will be used when the geth instance is launched.

Defaults to Self::DEFAULT_ADDR.

pub fn http_port<T: Into<u16>>(self, port: T) -> Self

Sets the --http.port which will be used when the geth instance is launched.

Defaults to Self::DEFAULT_HTTP_PORT.

pub fn http_corsdomain<T: Into<String>>(self, cors: T) -> Self

Comma separated list of domains from which to accept cross origin requests.

pub fn http_vhosts<T: Into<String>>(self, vhosts: T) -> Self

Comma separated list of virtual hostnames from which to accept requests.

pub fn ws(self) -> Self

Whether to pass the --ws flag to the geth instance.

pub fn ws_api<T: Into<String>>(self, api: T) -> Self

The --ws.api which will be used when the geth instance is launched.

pub fn ws_addr<T: Into<String>>(self, addr: T) -> Self

Sets the --ws.addr which will be used when the geth instance is launched.

Defaults to Self::DEFAULT_ADDR.

pub fn ws_port<T: Into<u16>>(self, port: T) -> Self

Sets the --ws.port which will be used when the geth instance is launched.

Defaults to Self::DEFAULT_WS_PORT.

pub fn networkid(self, networkid: u64) -> Self

Sets the --networkid which will be used when the geth instance is launched.

pub fn datadir<T: Into<PathBuf>>(self, datadir: T) -> Self

Sets the --datadir which will be used when the geth instance is launched.

pub fn ephemeral_datadir(self) -> Self

Whether to keep the geth data directory.

pub fn verbosity(self, verbosity: u8) -> Self

The --verbosity level which will be used when the geth instance is launched.

pub const fn keep_stderr(self) -> Self

Keep the handle to geth’s stderr in order to read from it.

Caution: if the stderr handle isn’t used, this can end up blocking.

pub fn spawn(self) -> Result<ArkivInstance>

Consumes the builder and spawns an ArkivInstance.

By default, it’s expected that geth is on $PATH. The default behavior can be overriden with Arkiv::program, or Arkiv::fetch_tag. See process::Command::new for further details.