arkiv_sdk::providers

Trait StorageProvider

pub trait StorageProvider<S: StorageNetwork>:
    Provider<S>
    + Send
    + Sync {
    // Provided methods
    fn query(
        &self,
        query_expression: &str,
        options: QueryOpts,
    ) -> RpcCall<Value, Value> { ... }
    fn get_entity_count(&self) -> RpcCall<(), u64> { ... }
    fn get_number_of_used_slots(&self) -> RpcCall<(), U256> { ... }
    fn get_block_timing(&self) -> RpcCall<(), BlockTiming> { ... }
    fn storage_transaction(&self) -> S::StorageTransactionRequest { ... }
    fn send_storage_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: S::StorageTransactionRequest,
    ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn create_entities<'life0, 'async_trait>(
        &'life0 self,
        creates: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Create>,
    ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn update_entities<'life0, 'async_trait>(
        &'life0 self,
        updates: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Update>,
    ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn delete_entities<'life0, 'async_trait>(
        &'life0 self,
        deletes: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Delete>,
    ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn extend_entities<'life0, 'async_trait>(
        &'life0 self,
        extensions: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Extend>,
    ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn transfer_entities<'life0, 'async_trait>(
        &'life0 self,
        transfers: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Chown>,
    ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn subscribe_storage_events<'life0, 'async_trait>(
        &'life0 self,
        f: fn(Filter) -> Filter,
    ) -> Pin<Box<dyn Future<Output = TransportResult<Map<SubscriptionStream<Log>, fn(Log) -> Result<S::StorageEvent, <S::StorageEvent as TryFrom<Log>>::Error>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

Adds storage methods to a alloy::providers::Provider.

alloy::network Documentation

This trait maintains dynamic compatibility with alloy::providers::DynProvider.

Provided Methods

fn query( &self, query_expression: &str, options: QueryOpts, ) -> RpcCall<Value, Value>

Filter and retrieve entities matching a query expression.

TODO: When we make a query, the returned value from the request is still JSON, however, this may be less idiomatic for practical use if the end user must do non-trivial work to get a decompressed, decoded and deserialized payload.

fn get_entity_count(&self) -> RpcCall<(), u64>

Returns total number of entities at current block.

fn get_number_of_used_slots(&self) -> RpcCall<(), U256>

Returns storage slot usage accounting.

fn get_block_timing(&self) -> RpcCall<(), BlockTiming>

Returns current block timing information.

fn storage_transaction(&self) -> S::StorageTransactionRequest

Construct the default StorageNetwork::StorageTransactionRequest.

fn send_storage_transaction<'life0, 'async_trait>( &'life0 self, tx: S::StorageTransactionRequest, ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn create_entities<'life0, 'async_trait>( &'life0 self, creates: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Create>, ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn update_entities<'life0, 'async_trait>( &'life0 self, updates: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Update>, ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn delete_entities<'life0, 'async_trait>( &'life0 self, deletes: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Delete>, ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn extend_entities<'life0, 'async_trait>( &'life0 self, extensions: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Extend>, ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn transfer_entities<'life0, 'async_trait>( &'life0 self, transfers: Vec<<<S::StorageTransactionRequest as StorageTransactionBuilder<S>>::Payload as PayloadBuilder<S>>::Chown>, ) -> Pin<Box<dyn Future<Output = TransportResult<PendingTransactionBuilder<S>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn subscribe_storage_events<'life0, 'async_trait>( &'life0 self, f: fn(Filter) -> Filter, ) -> Pin<Box<dyn Future<Output = TransportResult<Map<SubscriptionStream<Log>, fn(Log) -> Result<S::StorageEvent, <S::StorageEvent as TryFrom<Log>>::Error>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

A convenience method for subscribing to a stream of events from the StorageNetwork’s storage contract. Provides a closure over a [alloy::rpc::types::Filter] with the crate::tx::StorageTransactionRequest::STORAGE_ADDRESS pre-populated and attempts to convert the [alloy::sol] contract types into crate::network::StorageNetwork::StorageEvent.