arkiv_sdk

Trait Provider

pub trait Provider<N = Ethereum>: Send + Sync
where
    N: Network,
{

    // Required method
    fn root(&self) -> &RootProvider<N>;

    // Provided methods
    fn builder() -> ProviderBuilder<Identity, Identity, N>
       where Self: Sized { ... }
    fn client(&self) -> &RpcClientInner { ... }
    fn weak_client(&self) -> Weak<RpcClientInner> { ... }
    fn erased(self) -> DynProvider<N>
       where Self: Sized + 'static { ... }
    fn get_accounts(&self) -> ProviderCall<[(); 0], Vec<Address>> { ... }
    fn get_blob_base_fee(&self) -> ProviderCall<[(); 0], Uint<128, 2>, u128> { ... }
    fn get_block_number(&self) -> ProviderCall<[(); 0], Uint<64, 1>, u64> { ... }
    fn get_block_number_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn call(&self, tx: <N as Network>::TransactionRequest) -> EthCall<N, Bytes> { ... }
    fn call_many<'req>(
        &self,
        bundles: &'req [Bundle],
    ) -> EthCallMany<'req, N, Vec<Vec<EthCallResponse>>> { ... }
    fn multicall(&self) -> MulticallBuilder<Empty, &Self, N>
       where Self: Sized { ... }
    fn simulate<'req>(
        &self,
        payload: &'req SimulatePayload,
    ) -> RpcWithBlock<&'req SimulatePayload, Vec<SimulatedBlock<<N as Network>::BlockResponse>>> { ... }
    fn get_chain_id(&self) -> ProviderCall<[(); 0], Uint<64, 1>, u64> { ... }
    fn create_access_list<'a>(
        &self,
        request: &'a <N as Network>::TransactionRequest,
    ) -> RpcWithBlock<&'a <N as Network>::TransactionRequest, AccessListResult> { ... }
    fn estimate_gas(
        &self,
        tx: <N as Network>::TransactionRequest,
    ) -> EthCall<N, Uint<64, 1>, u64> { ... }
    fn estimate_eip1559_fees_with<'life0, 'async_trait>(
        &'life0 self,
        estimator: Eip1559Estimator,
    ) -> Pin<Box<dyn Future<Output = Result<Eip1559Estimation, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn estimate_eip1559_fees<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Eip1559Estimation, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_fee_history<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_count: u64,
        last_block: BlockNumberOrTag,
        reward_percentiles: &'life1 [f64],
    ) -> Pin<Box<dyn Future<Output = Result<FeeHistory, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn get_gas_price(&self) -> ProviderCall<[(); 0], Uint<128, 2>, u128> { ... }
    fn get_account_info(
        &self,
        address: Address,
    ) -> RpcWithBlock<Address, AccountInfo> { ... }
    fn get_account(
        &self,
        address: Address,
    ) -> RpcWithBlock<Address, TrieAccount> { ... }
    fn get_balance(
        &self,
        address: Address,
    ) -> RpcWithBlock<Address, Uint<256, 4>> { ... }
    fn get_block(
        &self,
        block: BlockId,
    ) -> EthGetBlock<<N as Network>::BlockResponse> { ... }
    fn get_block_by_hash(
        &self,
        hash: FixedBytes<32>,
    ) -> EthGetBlock<<N as Network>::BlockResponse> { ... }
    fn get_block_by_number(
        &self,
        number: BlockNumberOrTag,
    ) -> EthGetBlock<<N as Network>::BlockResponse> { ... }
    fn get_block_transaction_count_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: FixedBytes<32>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_block_transaction_count_by_number<'life0, 'async_trait>(
        &'life0 self,
        block_number: BlockNumberOrTag,
    ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_block_receipts(
        &self,
        block: BlockId,
    ) -> ProviderCall<(BlockId,), Option<Vec<<N as Network>::ReceiptResponse>>> { ... }
    fn get_code_at(&self, address: Address) -> RpcWithBlock<Address, Bytes> { ... }
    fn watch_blocks<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<FixedBytes<32>>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn watch_full_blocks<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<WatchBlocks<<N as Network>::BlockResponse>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn watch_pending_transactions<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<FixedBytes<32>>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn watch_logs<'life0, 'life1, 'async_trait>(
        &'life0 self,
        filter: &'life1 Filter,
    ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<Log>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn watch_full_pending_transactions<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<<N as Network>::TransactionResponse>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_filter_changes<'life0, 'async_trait, R>(
        &'life0 self,
        id: Uint<256, 4>,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<R>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: Sized + 'async_trait,
             R: 'async_trait + RpcRecv { ... }
    fn get_filter_changes_dyn<'life0, 'async_trait>(
        &'life0 self,
        id: Uint<256, 4>,
    ) -> Pin<Box<dyn Future<Output = Result<FilterChanges, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_filter_logs<'life0, 'async_trait>(
        &'life0 self,
        id: Uint<256, 4>,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Log>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn uninstall_filter<'life0, 'async_trait>(
        &'life0 self,
        id: Uint<256, 4>,
    ) -> Pin<Box<dyn Future<Output = Result<bool, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn watch_pending_transaction<'life0, 'async_trait>(
        &'life0 self,
        config: PendingTransactionConfig,
    ) -> Pin<Box<dyn Future<Output = Result<PendingTransaction, PendingTransactionError>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_logs<'life0, 'life1, 'async_trait>(
        &'life0 self,
        filter: &'life1 Filter,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Log>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn get_proof(
        &self,
        address: Address,
        keys: Vec<FixedBytes<32>>,
    ) -> RpcWithBlock<(Address, Vec<FixedBytes<32>>), EIP1186AccountProofResponse> { ... }
    fn get_storage_at(
        &self,
        address: Address,
        key: Uint<256, 4>,
    ) -> RpcWithBlock<(Address, Uint<256, 4>), Uint<256, 4>> { ... }
    fn get_transaction_by_sender_nonce(
        &self,
        sender: Address,
        nonce: u64,
    ) -> ProviderCall<(Address, Uint<64, 1>), Option<<N as Network>::TransactionResponse>> { ... }
    fn get_transaction_by_hash(
        &self,
        hash: FixedBytes<32>,
    ) -> ProviderCall<(FixedBytes<32>,), Option<<N as Network>::TransactionResponse>> { ... }
    fn get_transaction_by_block_hash_and_index(
        &self,
        block_hash: FixedBytes<32>,
        index: usize,
    ) -> ProviderCall<(FixedBytes<32>, Index), Option<<N as Network>::TransactionResponse>> { ... }
    fn get_raw_transaction_by_block_hash_and_index(
        &self,
        block_hash: FixedBytes<32>,
        index: usize,
    ) -> ProviderCall<(FixedBytes<32>, Index), Option<Bytes>> { ... }
    fn get_transaction_by_block_number_and_index(
        &self,
        block_number: BlockNumberOrTag,
        index: usize,
    ) -> ProviderCall<(BlockNumberOrTag, Index), Option<<N as Network>::TransactionResponse>> { ... }
    fn get_raw_transaction_by_block_number_and_index(
        &self,
        block_number: BlockNumberOrTag,
        index: usize,
    ) -> ProviderCall<(BlockNumberOrTag, Index), Option<Bytes>> { ... }
    fn get_raw_transaction_by_hash(
        &self,
        hash: FixedBytes<32>,
    ) -> ProviderCall<(FixedBytes<32>,), Option<Bytes>> { ... }
    fn get_transaction_count(
        &self,
        address: Address,
    ) -> RpcWithBlock<Address, Uint<64, 1>, u64> { ... }
    fn get_transaction_receipt(
        &self,
        hash: FixedBytes<32>,
    ) -> ProviderCall<(FixedBytes<32>,), Option<<N as Network>::ReceiptResponse>> { ... }
    fn get_uncle<'life0, 'async_trait>(
        &'life0 self,
        tag: BlockId,
        idx: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Option<<N as Network>::BlockResponse>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_uncle_count<'life0, 'async_trait>(
        &'life0 self,
        tag: BlockId,
    ) -> Pin<Box<dyn Future<Output = Result<u64, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn get_max_priority_fee_per_gas(
        &self,
    ) -> ProviderCall<[(); 0], Uint<128, 2>, u128> { ... }
    fn new_block_filter<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Uint<256, 4>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn new_filter<'life0, 'life1, 'async_trait>(
        &'life0 self,
        filter: &'life1 Filter,
    ) -> Pin<Box<dyn Future<Output = Result<Uint<256, 4>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn new_pending_transactions_filter<'life0, 'async_trait>(
        &'life0 self,
        full: bool,
    ) -> Pin<Box<dyn Future<Output = Result<Uint<256, 4>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn send_raw_transaction<'life0, 'life1, 'async_trait>(
        &'life0 self,
        encoded_tx: &'life1 [u8],
    ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn send_raw_transaction_sync<'life0, 'life1, 'async_trait>(
        &'life0 self,
        encoded_tx: &'life1 [u8],
    ) -> Pin<Box<dyn Future<Output = Result<<N as Network>::ReceiptResponse, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn send_raw_transaction_conditional<'life0, 'life1, 'async_trait>(
        &'life0 self,
        encoded_tx: &'life1 [u8],
        conditional: TransactionConditional,
    ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn send_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: <N as Network>::TransactionRequest,
    ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn send_tx_envelope<'life0, 'async_trait>(
        &'life0 self,
        tx: <N as Network>::TxEnvelope,
    ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn send_transaction_sync<'life0, 'async_trait>(
        &'life0 self,
        tx: <N as Network>::TransactionRequest,
    ) -> Pin<Box<dyn Future<Output = Result<<N as Network>::ReceiptResponse, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn sign_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: <N as Network>::TransactionRequest,
    ) -> Pin<Box<dyn Future<Output = Result<Bytes, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn subscribe_blocks(
        &self,
    ) -> GetSubscription<(SubscriptionKind,), <N as Network>::HeaderResponse> { ... }
    fn subscribe_full_blocks(&self) -> SubFullBlocks<N> { ... }
    fn subscribe_pending_transactions(
        &self,
    ) -> GetSubscription<(SubscriptionKind,), FixedBytes<32>> { ... }
    fn subscribe_full_pending_transactions(
        &self,
    ) -> GetSubscription<(SubscriptionKind, Params), <N as Network>::TransactionResponse> { ... }
    fn subscribe_logs(
        &self,
        filter: &Filter,
    ) -> GetSubscription<(SubscriptionKind, Params), Log> { ... }
    fn subscribe<P, R>(&self, params: P) -> GetSubscription<P, R>
       where P: RpcSend,
             R: RpcRecv,
             Self: Sized { ... }
    fn subscribe_to<R>(
        &self,
        method: &'static str,
    ) -> GetSubscription<[(); 0], R>
       where R: RpcRecv,
             Self: Sized { ... }
    fn unsubscribe<'life0, 'async_trait>(
        &'life0 self,
        id: FixedBytes<32>,
    ) -> Pin<Box<dyn Future<Output = Result<(), RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait { ... }
    fn syncing(&self) -> ProviderCall<[(); 0], SyncStatus> { ... }
    fn get_client_version(&self) -> ProviderCall<[(); 0], String> { ... }
    fn get_sha3(&self, data: &[u8]) -> ProviderCall<(String,), FixedBytes<32>> { ... }
    fn get_net_version(&self) -> ProviderCall<[(); 0], Uint<64, 1>, u64> { ... }
    fn raw_request<'life0, 'async_trait, P, R>(
        &'life0 self,
        method: Cow<'static, str>,
        params: P,
    ) -> Pin<Box<dyn Future<Output = Result<R, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             P: RpcSend + 'async_trait,
             R: RpcRecv + 'async_trait,
             Self: Sized + 'async_trait { ... }
    fn raw_request_dyn<'life0, 'life1, 'async_trait>(
        &'life0 self,
        method: Cow<'static, str>,
        params: &'life1 RawValue,
    ) -> Pin<Box<dyn Future<Output = Result<Box<RawValue>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             'life1: 'async_trait,
             Self: 'async_trait { ... }
    fn transaction_request(&self) -> <N as Network>::TransactionRequest { ... }

}

Re-export of commonly used types from alloy. Ethereum JSON-RPC interface.

Subscriptions

The provider supports pubsub subscriptions to new block headers and pending transactions. This is only available on pubsub clients, such as Websockets or IPC.

For a polling alternatives available over HTTP, use the watch_* methods. However, be aware that polling increases RPC usage drastically.

Special treatment of EIP-1559

While many RPC features are encapsulated by extension traits, EIP-1559 fee estimation is generally assumed to be on by default. We generally assume that EIP-1559 is supported by the client and will proactively use it by default.

As a result, the provider supports EIP-1559 fee estimation the ethereum TransactionBuilder will use it by default. We acknowledge that this means EIP-1559 has a privileged status in comparison to other transaction types. Networks that DO NOT support EIP-1559 should create their own TransactionBuilder and Fillers to change this behavior.

Required Methods

fn root(&self) -> &RootProvider<N>

Returns the root provider.

Provided Methods

fn builder() -> ProviderBuilder<Identity, Identity, N>
where Self: Sized,

Returns the ProviderBuilder to build on.

fn client(&self) -> &RpcClientInner

Returns the RPC client used to send requests.

NOTE: this method should not be overridden.

fn weak_client(&self) -> Weak<RpcClientInner>

Returns a Weak RPC client used to send requests.

NOTE: this method should not be overridden.

fn erased(self) -> DynProvider<N>
where Self: Sized + 'static,

Returns a type erased provider wrapped in Arc. See DynProvider.

use alloy_provider::{DynProvider, Provider, ProviderBuilder};

let provider: DynProvider =
    ProviderBuilder::new().on_builtin("http://localhost:8080").await?.erased();
let block = provider.get_block_number().await?;

fn get_accounts(&self) -> ProviderCall<[(); 0], Vec<Address>>

Gets the accounts in the remote node. This is usually empty unless you’re using a local node.

fn get_blob_base_fee(&self) -> ProviderCall<[(); 0], Uint<128, 2>, u128>

Returns the base fee per blob gas (blob gas price) in wei.

fn get_block_number(&self) -> ProviderCall<[(); 0], Uint<64, 1>, u64>

Get the last block number available.

fn get_block_number_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Get the block number for a given block identifier.

This is a convenience function that fetches the full block when the block identifier is not a number.

fn call(&self, tx: <N as Network>::TransactionRequest) -> EthCall<N, Bytes>

Execute a smart contract call with a transaction request and state overrides, without publishing a transaction.

This function returns [EthCall] which can be used to execute the call, or to add StateOverride or a [BlockId]. If no overrides or block ID is provided, the call will be executed on the pending block with the current state.

Examples

// Execute a call on the latest block, with no state overrides
let output = provider.call(tx).await?;

fn call_many<'req>( &self, bundles: &'req [Bundle], ) -> EthCallMany<'req, N, Vec<Vec<EthCallResponse>>>

Execute a list of [Bundle] against the provided StateContext and StateOverride, without publishing a transaction.

This function returns [EthCallMany] builder which is used to execute the call, and also set the StateContext and StateOverride.

fn multicall(&self) -> MulticallBuilder<Empty, &Self, N>
where Self: Sized,

Execute a multicall by leveraging the [MulticallBuilder].

Call [MulticallBuilder::dynamic] to add calls dynamically instead.

See the [MulticallBuilder] documentation for more details.

fn simulate<'req>( &self, payload: &'req SimulatePayload, ) -> RpcWithBlock<&'req SimulatePayload, Vec<SimulatedBlock<<N as Network>::BlockResponse>>>

Executes an arbitrary number of transactions on top of the requested state.

The transactions are packed into individual blocks. Overrides can be provided.

fn get_chain_id(&self) -> ProviderCall<[(); 0], Uint<64, 1>, u64>

Gets the chain ID.

fn create_access_list<'a>( &self, request: &'a <N as Network>::TransactionRequest, ) -> RpcWithBlock<&'a <N as Network>::TransactionRequest, AccessListResult>

Create an EIP-2930 access list.

fn estimate_gas( &self, tx: <N as Network>::TransactionRequest, ) -> EthCall<N, Uint<64, 1>, u64>

Create an [EthCall] future to estimate the gas required for a transaction.

The future can be used to specify a StateOverride or [BlockId] before dispatching the call. If no overrides or block ID is provided, the gas estimate will be computed for the pending block with the current state.

Note

Not all client implementations support state overrides for eth_estimateGas.

fn estimate_eip1559_fees_with<'life0, 'async_trait>( &'life0 self, estimator: Eip1559Estimator, ) -> Pin<Box<dyn Future<Output = Result<Eip1559Estimation, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Estimates the EIP1559 maxFeePerGas and maxPriorityFeePerGas fields.

Receives an [Eip1559Estimator] that can be used to modify how to estimate these fees.

fn estimate_eip1559_fees<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Eip1559Estimation, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Estimates the EIP1559 maxFeePerGas and maxPriorityFeePerGas fields.

Uses the builtin estimator [utils::eip1559_default_estimator] function.

fn get_fee_history<'life0, 'life1, 'async_trait>( &'life0 self, block_count: u64, last_block: BlockNumberOrTag, reward_percentiles: &'life1 [f64], ) -> Pin<Box<dyn Future<Output = Result<FeeHistory, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Returns a collection of historical gas information [FeeHistory] which can be used to calculate the EIP1559 fields maxFeePerGas and maxPriorityFeePerGas. block_count can range from 1 to 1024 blocks in a single request.

fn get_gas_price(&self) -> ProviderCall<[(); 0], Uint<128, 2>, u128>

Gets the current gas price in wei.

fn get_account_info( &self, address: Address, ) -> RpcWithBlock<Address, AccountInfo>

Retrieves account information (Account) for the given Address at the particular [BlockId].

Note: This is slightly different than eth_getAccount and not all clients support this endpoint.

fn get_account(&self, address: Address) -> RpcWithBlock<Address, TrieAccount>

Retrieves account information (Account) for the given Address at the particular [BlockId].

fn get_balance(&self, address: Address) -> RpcWithBlock<Address, Uint<256, 4>>

Gets the balance of the account.

Defaults to the latest block. See also [RpcWithBlock::block_id].

fn get_block( &self, block: BlockId, ) -> EthGetBlock<<N as Network>::BlockResponse>

Gets a block by either its hash, tag, or number

By default this fetches the block with only the transaction hashes, and not full transactions.

To get full transactions one can do:

let block = provider.get_block(BlockId::latest()).full().await.unwrap();

fn get_block_by_hash( &self, hash: FixedBytes<32>, ) -> EthGetBlock<<N as Network>::BlockResponse>

Gets a block by its [BlockHash]

By default this fetches the block with only the transaction hashes populated in the block, and not the full transactions.

Examples

#[tokio::main]
async fn main() {
    let provider =
        ProviderBuilder::new().connect_http("https://eth.merkle.io".parse().unwrap());
    let block_hash = b256!("6032d03ee8e43e8999c2943152a4daebfc4b75b7f7a9647d2677299d215127da");

    // Gets a block by its hash with only transactions hashes.
    let block = provider.get_block_by_hash(block_hash).await.unwrap();

    // Gets a block by its hash with full transactions.
    let block = provider.get_block_by_hash(block_hash).full().await.unwrap();
}

fn get_block_by_number( &self, number: BlockNumberOrTag, ) -> EthGetBlock<<N as Network>::BlockResponse>

Gets a block by its [BlockNumberOrTag]

By default this fetches the block with only the transaction hashes populated in the block, and not the full transactions.

Examples

#[tokio::main]
async fn main() {
    let provider =
        ProviderBuilder::new().connect_http("https://eth.merkle.io".parse().unwrap());
    let num = BlockNumberOrTag::Number(0);

    // Gets a block by its number with only transactions hashes.
    let block = provider.get_block_by_number(num).await.unwrap();

    // Gets a block by its number with full transactions.
    let block = provider.get_block_by_number(num).full().await.unwrap();
}

fn get_block_transaction_count_by_hash<'life0, 'async_trait>( &'life0 self, hash: FixedBytes<32>, ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Returns the number of transactions in a block from a block matching the given block hash.

fn get_block_transaction_count_by_number<'life0, 'async_trait>( &'life0 self, block_number: BlockNumberOrTag, ) -> Pin<Box<dyn Future<Output = Result<Option<u64>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Returns the number of transactions in a block matching the given block number.

fn get_block_receipts( &self, block: BlockId, ) -> ProviderCall<(BlockId,), Option<Vec<<N as Network>::ReceiptResponse>>>

Gets the selected block [BlockId] receipts.

fn get_code_at(&self, address: Address) -> RpcWithBlock<Address, Bytes>

Gets the bytecode located at the corresponding Address.

fn watch_blocks<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<FixedBytes<32>>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Watch for new blocks by polling the provider with eth_getFilterChanges.

Returns a builder that is used to configure the poller. See [PollerBuilder] for more details.

Examples

Get the next 5 blocks:

use futures::StreamExt;

let poller = provider.watch_blocks().await?;
let mut stream = poller.into_stream().flat_map(futures::stream::iter).take(5);
while let Some(block_hash) = stream.next().await {
   println!("new block: {block_hash}");
}

fn watch_full_blocks<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<WatchBlocks<<N as Network>::BlockResponse>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Watch for new blocks by polling the provider with eth_getFilterChanges and transforming the returned block hashes into full blocks bodies.

Returns the [WatchBlocks] type which consumes the stream of block hashes from [PollerBuilder] and returns a stream of [BlockResponse]’s.

Examples

Get the next 5 full blocks:

use futures::StreamExt;

let poller = provider.watch_full_blocks().await?.full();
let mut stream = poller.into_stream().flat_map(futures::stream::iter).take(5);
while let Some(block) = stream.next().await {
  println!("new block: {block:#?}");
}

fn watch_pending_transactions<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<FixedBytes<32>>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Watch for new pending transaction by polling the provider with eth_getFilterChanges.

Returns a builder that is used to configure the poller. See [PollerBuilder] for more details.

Examples

Get the next 5 pending transaction hashes:

use futures::StreamExt;

let poller = provider.watch_pending_transactions().await?;
let mut stream = poller.into_stream().flat_map(futures::stream::iter).take(5);
while let Some(tx_hash) = stream.next().await {
   println!("new pending transaction hash: {tx_hash}");
}

fn watch_logs<'life0, 'life1, 'async_trait>( &'life0 self, filter: &'life1 Filter, ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<Log>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Watch for new logs using the given filter by polling the provider with eth_getFilterChanges.

Returns a builder that is used to configure the poller. See [PollerBuilder] for more details.

Examples

Get the next 5 USDC transfer logs:

use alloy_primitives::{address, b256};
use alloy_rpc_types_eth::Filter;
use futures::StreamExt;

let address = address!("a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48");
let transfer_signature = b256!("ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef");
let filter = Filter::new().address(address).event_signature(transfer_signature);

let poller = provider.watch_logs(&filter).await?;
let mut stream = poller.into_stream().flat_map(futures::stream::iter).take(5);
while let Some(log) = stream.next().await {
   println!("new log: {log:#?}");
}

fn watch_full_pending_transactions<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<PollerBuilder<(Uint<256, 4>,), Vec<<N as Network>::TransactionResponse>>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Watch for new pending transaction bodies by polling the provider with eth_getFilterChanges.

Returns a builder that is used to configure the poller. See [PollerBuilder] for more details.

Support

This endpoint might not be supported by all clients.

Examples

Get the next 5 pending transaction bodies:

use futures::StreamExt;

let poller = provider.watch_full_pending_transactions().await?;
let mut stream = poller.into_stream().flat_map(futures::stream::iter).take(5);
while let Some(tx) = stream.next().await {
   println!("new pending transaction: {tx:#?}");
}

fn get_filter_changes<'life0, 'async_trait, R>( &'life0 self, id: Uint<256, 4>, ) -> Pin<Box<dyn Future<Output = Result<Vec<R>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: Sized + 'async_trait, R: 'async_trait + RpcRecv,

Get a list of values that have been added since the last poll.

The return value depends on what stream id corresponds to. See [FilterChanges] for all possible return values.

fn get_filter_changes_dyn<'life0, 'async_trait>( &'life0 self, id: Uint<256, 4>, ) -> Pin<Box<dyn Future<Output = Result<FilterChanges, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Get a list of values that have been added since the last poll.

This returns an enum over all possible return values. You probably want to use get_filter_changes instead.

fn get_filter_logs<'life0, 'async_trait>( &'life0 self, id: Uint<256, 4>, ) -> Pin<Box<dyn Future<Output = Result<Vec<Log>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Retrieves a Vec<Log> for the given filter ID.

fn uninstall_filter<'life0, 'async_trait>( &'life0 self, id: Uint<256, 4>, ) -> Pin<Box<dyn Future<Output = Result<bool, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Request provider to uninstall the filter with the given ID.

fn watch_pending_transaction<'life0, 'async_trait>( &'life0 self, config: PendingTransactionConfig, ) -> Pin<Box<dyn Future<Output = Result<PendingTransaction, PendingTransactionError>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Watch for the confirmation of a single pending transaction with the given configuration.

Note that this is handled internally rather than calling any specific RPC method, and as such should not be overridden.

fn get_logs<'life0, 'life1, 'async_trait>( &'life0 self, filter: &'life1 Filter, ) -> Pin<Box<dyn Future<Output = Result<Vec<Log>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Retrieves a Vec<Log> with the given [Filter].

fn get_proof( &self, address: Address, keys: Vec<FixedBytes<32>>, ) -> RpcWithBlock<(Address, Vec<FixedBytes<32>>), EIP1186AccountProofResponse>

Get the account and storage values of the specified account including the merkle proofs.

This call can be used to verify that the data has not been tampered with.

fn get_storage_at( &self, address: Address, key: Uint<256, 4>, ) -> RpcWithBlock<(Address, Uint<256, 4>), Uint<256, 4>>

Gets the specified storage value from Address.

fn get_transaction_by_sender_nonce( &self, sender: Address, nonce: u64, ) -> ProviderCall<(Address, Uint<64, 1>), Option<<N as Network>::TransactionResponse>>

Gets a transaction by its sender and nonce.

Note: not supported by all clients.

fn get_transaction_by_hash( &self, hash: FixedBytes<32>, ) -> ProviderCall<(FixedBytes<32>,), Option<<N as Network>::TransactionResponse>>

Gets a transaction by its [TxHash].

fn get_transaction_by_block_hash_and_index( &self, block_hash: FixedBytes<32>, index: usize, ) -> ProviderCall<(FixedBytes<32>, Index), Option<<N as Network>::TransactionResponse>>

Gets a transaction by block hash and transaction index position.

fn get_raw_transaction_by_block_hash_and_index( &self, block_hash: FixedBytes<32>, index: usize, ) -> ProviderCall<(FixedBytes<32>, Index), Option<Bytes>>

Gets a raw transaction by block hash and transaction index position.

fn get_transaction_by_block_number_and_index( &self, block_number: BlockNumberOrTag, index: usize, ) -> ProviderCall<(BlockNumberOrTag, Index), Option<<N as Network>::TransactionResponse>>

Gets a transaction by block number and transaction index position.

fn get_raw_transaction_by_block_number_and_index( &self, block_number: BlockNumberOrTag, index: usize, ) -> ProviderCall<(BlockNumberOrTag, Index), Option<Bytes>>

Gets a raw transaction by block number and transaction index position.

fn get_raw_transaction_by_hash( &self, hash: FixedBytes<32>, ) -> ProviderCall<(FixedBytes<32>,), Option<Bytes>>

Returns the EIP-2718 encoded transaction if it exists, see also Decodable2718.

If the transaction is an EIP-4844 transaction that is still in the pool (pending) it will include the sidecar, otherwise it will the consensus variant without the sidecar: TxEip4844.

This can be decoded into TxEnvelope.

fn get_transaction_count( &self, address: Address, ) -> RpcWithBlock<Address, Uint<64, 1>, u64>

Gets the transaction count (AKA “nonce”) of the corresponding address.

fn get_transaction_receipt( &self, hash: FixedBytes<32>, ) -> ProviderCall<(FixedBytes<32>,), Option<<N as Network>::ReceiptResponse>>

Gets a transaction receipt if it exists, by its [TxHash].

fn get_uncle<'life0, 'async_trait>( &'life0 self, tag: BlockId, idx: u64, ) -> Pin<Box<dyn Future<Output = Result<Option<<N as Network>::BlockResponse>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Gets an uncle block through the tag [BlockId] and index u64.

fn get_uncle_count<'life0, 'async_trait>( &'life0 self, tag: BlockId, ) -> Pin<Box<dyn Future<Output = Result<u64, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Gets the number of uncles for the block specified by the tag [BlockId].

fn get_max_priority_fee_per_gas( &self, ) -> ProviderCall<[(); 0], Uint<128, 2>, u128>

Returns a suggestion for the current maxPriorityFeePerGas in wei.

fn new_block_filter<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Uint<256, 4>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Notify the provider that we are interested in new blocks.

Returns the ID to use with eth_getFilterChanges.

See also watch_blocks to configure a poller.

fn new_filter<'life0, 'life1, 'async_trait>( &'life0 self, filter: &'life1 Filter, ) -> Pin<Box<dyn Future<Output = Result<Uint<256, 4>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Notify the provider that we are interested in logs that match the given filter.

Returns the ID to use with eth_getFilterChanges.

See also watch_logs to configure a poller.

fn new_pending_transactions_filter<'life0, 'async_trait>( &'life0 self, full: bool, ) -> Pin<Box<dyn Future<Output = Result<Uint<256, 4>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Notify the provider that we are interested in new pending transactions.

If full is true, the stream will consist of full transaction bodies instead of just the hashes. This not supported by all clients.

Returns the ID to use with eth_getFilterChanges.

See also watch_pending_transactions to configure a poller.

fn send_raw_transaction<'life0, 'life1, 'async_trait>( &'life0 self, encoded_tx: &'life1 [u8], ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Broadcasts a raw transaction RLP bytes to the network.

See send_transaction for more details.

fn send_raw_transaction_sync<'life0, 'life1, 'async_trait>( &'life0 self, encoded_tx: &'life1 [u8], ) -> Pin<Box<dyn Future<Output = Result<<N as Network>::ReceiptResponse, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Broadcasts a raw transaction RLP bytes to the network and returns the transaction receipt after it has been mined.

Unlike send_raw_transaction which returns immediately with a transaction hash, this method waits on the server side until the transaction is included in a block and returns the receipt directly. This is an optimization that reduces the number of RPC calls needed to confirm a transaction.

This method implements the eth_sendRawTransactionSync RPC method as defined in EIP-7966.

Error Handling

If the transaction fails, you can extract the transaction hash from the error using [RpcError::tx_hash_data]:

match provider.send_raw_transaction_sync(encoded_tx).await {
    Ok(receipt) => {
        println!("Transaction successful: {}", receipt.transaction_hash());
    }
    Err(rpc_err) => {
        if let Some(tx_hash) = rpc_err.tx_hash_data() {
            println!("Transaction failed but hash available: {}", tx_hash);
        }
    }
}

Note: This is only available on certain clients that support the eth_sendRawTransactionSync RPC method, such as Anvil.

fn send_raw_transaction_conditional<'life0, 'life1, 'async_trait>( &'life0 self, encoded_tx: &'life1 [u8], conditional: TransactionConditional, ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Broadcasts a raw transaction RLP bytes with a conditional [TransactionConditional] to the network.

TransactionConditional represents the preconditions that determine the inclusion of the transaction, enforced out-of-protocol by the sequencer.

Note: This endpoint is only available on certain networks, e.g. opstack chains, polygon, bsc.

See [TransactionConditional] for more details.

fn send_transaction<'life0, 'async_trait>( &'life0 self, tx: <N as Network>::TransactionRequest, ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Broadcasts a transaction to the network.

Returns a [PendingTransactionBuilder] which can be used to configure how and when to await the transaction’s confirmation.

Examples

See PendingTransactionBuilder for more examples.

let tx_hash = provider.send_transaction(tx)
    .await?
    .with_required_confirmations(2)
    .with_timeout(Some(std::time::Duration::from_secs(60)))
    .watch()
    .await?;

fn send_tx_envelope<'life0, 'async_trait>( &'life0 self, tx: <N as Network>::TxEnvelope, ) -> Pin<Box<dyn Future<Output = Result<PendingTransactionBuilder<N>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Broadcasts a transaction envelope to the network.

Returns a [PendingTransactionBuilder] which can be used to configure how and when to await the transaction’s confirmation.

fn send_transaction_sync<'life0, 'async_trait>( &'life0 self, tx: <N as Network>::TransactionRequest, ) -> Pin<Box<dyn Future<Output = Result<<N as Network>::ReceiptResponse, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Sends a transaction and waits for its receipt in a single call.

This method combines transaction submission and receipt retrieval into a single async operation, providing a simpler API compared to the two-step process of send_transaction followed by waiting for confirmation.

Returns the transaction receipt directly after submission and confirmation.

Example

let receipt = provider.send_transaction_sync(tx).await?;
println!("Transaction hash: {}", receipt.transaction_hash());

Error Handling

If the transaction fails, you can extract the transaction hash from the error using [RpcError::tx_hash_data]:

match provider.send_transaction_sync(tx).await {
    Ok(receipt) => {
        println!("Transaction successful: {}", receipt.transaction_hash());
    }
    Err(rpc_err) => {
        if let Some(tx_hash) = rpc_err.tx_hash_data() {
            println!("Transaction failed but hash available: {}", tx_hash);
        }
    }
}

fn sign_transaction<'life0, 'async_trait>( &'life0 self, tx: <N as Network>::TransactionRequest, ) -> Pin<Box<dyn Future<Output = Result<Bytes, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Signs a transaction that can be submitted to the network later using Provider::send_raw_transaction.

The "eth_signTransaction" method is not supported by regular nodes.

fn subscribe_blocks( &self, ) -> GetSubscription<(SubscriptionKind,), <N as Network>::HeaderResponse>

Subscribe to a stream of new block headers.

Errors

This method is only available on pubsub clients, such as WebSockets or IPC, and will return a PubsubUnavailable transport error if the client does not support it.

For a polling alternative available over HTTP, use Provider::watch_blocks. However, be aware that polling increases RPC usage drastically.

Examples

use futures::StreamExt;

let sub = provider.subscribe_blocks().await?;
let mut stream = sub.into_stream().take(5);
while let Some(block) = stream.next().await {
   println!("new block: {block:#?}");
}

fn subscribe_full_blocks(&self) -> SubFullBlocks<N>

Subscribe to a stream of full block bodies.

Errors

This method is only available on pubsub clients, such as WebSockets or IPC, and will return a PubsubUnavailable transport error if the client does not support it.

Examples

use futures::StreamExt;

let sub = provider.subscribe_full_blocks().full().channel_size(10);
let mut stream = sub.into_stream().await?.take(5);

while let Some(block) = stream.next().await {
  println!("{block:#?}");
}

fn subscribe_pending_transactions( &self, ) -> GetSubscription<(SubscriptionKind,), FixedBytes<32>>

Subscribe to a stream of pending transaction hashes.

Errors

This method is only available on pubsub clients, such as WebSockets or IPC, and will return a PubsubUnavailable transport error if the client does not support it.

For a polling alternative available over HTTP, use Provider::watch_pending_transactions. However, be aware that polling increases RPC usage drastically.

Examples

use futures::StreamExt;

let sub = provider.subscribe_pending_transactions().await?;
let mut stream = sub.into_stream().take(5);
while let Some(tx_hash) = stream.next().await {
   println!("new pending transaction hash: {tx_hash}");
}

fn subscribe_full_pending_transactions( &self, ) -> GetSubscription<(SubscriptionKind, Params), <N as Network>::TransactionResponse>

Subscribe to a stream of pending transaction bodies.

Support

This endpoint is compatible only with Geth client version 1.11.0 or later.

Errors

This method is only available on pubsub clients, such as WebSockets or IPC, and will return a PubsubUnavailable transport error if the client does not support it.

For a polling alternative available over HTTP, use Provider::watch_full_pending_transactions. However, be aware that polling increases RPC usage drastically.

Examples

use futures::StreamExt;

let sub = provider.subscribe_full_pending_transactions().await?;
let mut stream = sub.into_stream().take(5);
while let Some(tx) = stream.next().await {
   println!("{tx:#?}");
}

fn subscribe_logs( &self, filter: &Filter, ) -> GetSubscription<(SubscriptionKind, Params), Log>

Subscribe to a stream of logs matching given filter.

Errors

This method is only available on pubsub clients, such as WebSockets or IPC, and will return a PubsubUnavailable transport error if the client does not support it.

For a polling alternative available over HTTP, use Provider::watch_logs. However, be aware that polling increases RPC usage drastically.

Examples

use futures::StreamExt;
use alloy_primitives::keccak256;
use alloy_rpc_types_eth::Filter;

let signature = keccak256("Transfer(address,address,uint256)".as_bytes());

let sub = provider.subscribe_logs(&Filter::new().event_signature(signature)).await?;
let mut stream = sub.into_stream().take(5);
while let Some(tx) = stream.next().await {
   println!("{tx:#?}");
}

fn subscribe<P, R>(&self, params: P) -> GetSubscription<P, R>
where P: RpcSend, R: RpcRecv, Self: Sized,

Subscribe to an RPC event.

fn subscribe_to<R>(&self, method: &'static str) -> GetSubscription<[(); 0], R>
where R: RpcRecv, Self: Sized,

Subscribe to a non-standard subscription method without parameters.

This is a helper method for creating subscriptions to methods that are not “eth_subscribe” and don’t require parameters. It automatically marks the request as a subscription.

Examples

use futures::StreamExt;

let sub = provider.subscribe_to::<alloy_rpc_types_admin::PeerEvent>("admin_peerEvents").await?;
let mut stream = sub.into_stream().take(5);
while let Some(event) = stream.next().await {
   println!("peer event: {event:#?}");
}

fn unsubscribe<'life0, 'async_trait>( &'life0 self, id: FixedBytes<32>, ) -> Pin<Box<dyn Future<Output = Result<(), RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Cancels a subscription given the subscription ID.

fn syncing(&self) -> ProviderCall<[(); 0], SyncStatus>

Gets syncing info.

fn get_client_version(&self) -> ProviderCall<[(); 0], String>

Gets the client version.

fn get_sha3(&self, data: &[u8]) -> ProviderCall<(String,), FixedBytes<32>>

Gets the Keccak-256 hash of the given data.

fn get_net_version(&self) -> ProviderCall<[(); 0], Uint<64, 1>, u64>

Gets the network ID. Same as eth_chainId.

fn raw_request<'life0, 'async_trait, P, R>( &'life0 self, method: Cow<'static, str>, params: P, ) -> Pin<Box<dyn Future<Output = Result<R, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, P: RpcSend + 'async_trait, R: RpcRecv + 'async_trait, Self: Sized + 'async_trait,

Sends a raw JSON-RPC request.

Examples

use alloy_rpc_types_eth::BlockNumberOrTag;
use alloy_rpc_client::NoParams;

// No parameters: `()`
let block_number = provider.raw_request("eth_blockNumber".into(), NoParams::default()).await?;

// One parameter: `(param,)` or `[param]`
let block = provider.raw_request("eth_getBlockByNumber".into(), (BlockNumberOrTag::Latest,)).await?;

// Two or more parameters: `(param1, param2, ...)` or `[param1, param2, ...]`
let full_block = provider.raw_request("eth_getBlockByNumber".into(), (BlockNumberOrTag::Latest, true)).await?;

fn raw_request_dyn<'life0, 'life1, 'async_trait>( &'life0 self, method: Cow<'static, str>, params: &'life1 RawValue, ) -> Pin<Box<dyn Future<Output = Result<Box<RawValue>, RpcError<TransportErrorKind>>> + Send + 'async_trait>>
where 'life0: 'async_trait, 'life1: 'async_trait, Self: 'async_trait,

Sends a raw JSON-RPC request with type-erased parameters and return.

Examples

use alloy_rpc_types_eth::BlockNumberOrTag;

// No parameters: `()`
let params = serde_json::value::to_raw_value(&())?;
let block_number = provider.raw_request_dyn("eth_blockNumber".into(), &params).await?;

// One parameter: `(param,)` or `[param]`
let params = serde_json::value::to_raw_value(&(BlockNumberOrTag::Latest,))?;
let block = provider.raw_request_dyn("eth_getBlockByNumber".into(), &params).await?;

// Two or more parameters: `(param1, param2, ...)` or `[param1, param2, ...]`
let params = serde_json::value::to_raw_value(&(BlockNumberOrTag::Latest, true))?;
let full_block = provider.raw_request_dyn("eth_getBlockByNumber".into(), &params).await?;

fn transaction_request(&self) -> <N as Network>::TransactionRequest

Creates a new TransactionRequest.