arkiv_sdk

Module tx

Re-exports

pub use receipt::ChownReceipt;

pub use receipt::CreateReceipt;

pub use receipt::DeleteReceipt;

pub use receipt::ExtendReceipt;

pub use receipt::TransactionReceipt;

pub use receipt::UpdateReceipt;

Modules

ops

receipt

Structs

StoragePayload

An RLP encodable storage transaction payload containing storage operations to be executed on the network.

StorageTransactionRequest

The underlying network transaction to be sent and an encodable payload containing storage operations. This type implements StorageTransactionBuilder and is intended to be used just as [alloy::network::Network::TransactionRequest].

Traits

PayloadBuilder

Update a mutable reference to a StorageTransactionBuilder payload.

StorageTransactionBuilder

Extension trait for adding storage functionality to [alloy::network::TransactionBuilder].