arkiv_sdk::tx::ops

Trait WithAttribute

pub trait WithAttribute<A> {
    // Required methods
    fn with_attribute(self, attribute: A) -> Self;
    fn extend_attributes<I>(self, iter: I) -> Self
       where I: IntoIterator<Item = A>;
}

A trait for attaching attributes to a transaction’s operation.

Implementors provide distinct behavior depending on the wrapper type used (e.g. string attributes vs numeric attributes).

Required Methods

fn with_attribute(self, attribute: A) -> Self

Add a single attribute to a transaction type.

fn extend_attributes<I>(self, iter: I) -> Self
where I: IntoIterator<Item = A>,

Extend attributes from any type which can produce an iterator. This can be particularly useful for avoiding duplicate keys using std::collections::HashMap.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.