arkiv_sdk::tx

Trait StorageTransactionBuilder

pub trait StorageTransactionBuilder<S: StorageNetwork>:
    Default
    + Deref<Target = S::TransactionRequest>
    + DerefMut
    + Sized {
    type Payload: Default + Encodable + PayloadBuilder<S> + Send;

    const STORAGE_ADDRESS: Address;

    // Required methods
    fn payload(&self) -> &Self::Payload;
    fn payload_mut(&mut self) -> &mut Self::Payload;

    // Provided methods
    fn create_entities(
        self,
        creates: Vec<<Self::Payload as PayloadBuilder<S>>::Create>,
    ) -> Self { ... }
    fn update_entities(
        self,
        updates: Vec<<Self::Payload as PayloadBuilder<S>>::Update>,
    ) -> Self { ... }
    fn delete_entities(
        self,
        deletes: Vec<<Self::Payload as PayloadBuilder<S>>::Delete>,
    ) -> Self { ... }
    fn extend_entities(
        self,
        extensions: Vec<<Self::Payload as PayloadBuilder<S>>::Extend>,
    ) -> Self { ... }
    fn transfer_entities(
        self,
        transfers: Vec<<Self::Payload as PayloadBuilder<S>>::Chown>,
    ) -> Self { ... }
    fn into_request(self) -> S::TransactionRequest { ... }
}

Extension trait for adding storage functionality to [alloy::network::TransactionBuilder].

Required Associated Constants

const STORAGE_ADDRESS: Address

The address of the storage contract.

Required Associated Types

type Payload: Default + Encodable + PayloadBuilder<S> + Send

Required Methods

fn payload(&self) -> &Self::Payload

Access the underlying payload

fn payload_mut(&mut self) -> &mut Self::Payload

Provided Methods

fn create_entities( self, creates: Vec<<Self::Payload as PayloadBuilder<S>>::Create>, ) -> Self

Add a list of Create operations to the encodable transaction inputs.

fn update_entities( self, updates: Vec<<Self::Payload as PayloadBuilder<S>>::Update>, ) -> Self

Add a list of Update operations to the encodable transaction inputs.

fn delete_entities( self, deletes: Vec<<Self::Payload as PayloadBuilder<S>>::Delete>, ) -> Self

Add a list of Delete operations to the encodable transaction inputs.

fn extend_entities( self, extensions: Vec<<Self::Payload as PayloadBuilder<S>>::Extend>, ) -> Self

Add a list of Extend operations to the encodable transaction inputs.

fn transfer_entities( self, transfers: Vec<<Self::Payload as PayloadBuilder<S>>::Chown>, ) -> Self

Add a list of Chown operations to the encodable transaction inputs.

fn into_request(self) -> S::TransactionRequest

Encode and compress the storage payload and set the transaction input and [alloy::primitives::TxKind], returning the inner [alloy::network::Network::TransactionRequest].

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.